rename: AdminEditAccount 컴포넌트의 template 상위 폴더 생성
